+++ RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 # Release notes
 
+## v0.6.0 (2025-06-03)
+
+- On scan, APIC data is now extracted in directories which do not
+  contain `cover.jpg`
+- In 0.5.0, ArtistCutoff was added to the config file, but the value
+  was not used. This release corrects that error.
+- All ID3 related code removed from Catalog
+
 ## v0.5.0 (2025-05-31)
 
 - Multiple `tmctool` scan/metadata updates